feat(legal-gazette): UI improvements and trpc integration (#696)
* Changing out api calls for trpc

* feat(legal-gazette): Refactor advert details and add accordion components

* Updating the html structure of the components previously implemented

* feat(legal-gazette): Update TRPC API calls and refactor advert-related components

* custom hooks for each field update in advert - getting started

* Advert update fixes

* updating trpc integration

* publications and moving getting pageparams on page level and more

* thissucks

* Fixing legacyhtml

* adding court and judgement fields v1

* judgement and court district works now

* remove client-components folder

* move accordion items to field set and update divisonmeeting from accordionitem to fields

* settlement getting started

* settlement fields setup - missing trpc integration

* The API now return if the current user can edit the advert.

* TRPC integration for settlements

* refactor: streamline advert publishing logic and integrate TRPC for publishAdverts

* Removing todo comments

* Updating update hooks

---------

+++ apps/legal-gazette-api/src/modules/advert-publications/advert-publication.service.ts
@@ -121,6 +121,14 @@ export class AdvertPublicationService implements IAdvertPublicationService {
   }
 
   async deleteAdvertPublication(id: string, pubId: string): Promise<void> {
+    const count = await this.advertPublicationModel.count({
+      where: { advertId: id },
+    })
+
+    if (count <= 1) {
+      throw new BadRequestException('At least one publication must remain')
+    }
+
     await this.advertPublicationModel.destroy({
       where: {
         id: pubId,
@@ -136,10 +144,6 @@ export class AdvertPublicationService implements IAdvertPublicationService {
         ['publishedAt', 'ASC'],
       ],
     })
-
-    if (publications.length <= 1) {
-      throw new BadRequestException('At least one publication must remain')
-    }
 
     publications.forEach(async (publication, index) => {
       await publication.update({ versionNumber: index + 1 })
